chore: add revalidate guard and client component
